third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animations/setting-the-timeline-of-an-animation-expected.txt

This is a testharness.js-based test.
FAIL After setting timeline on paused animation it is still paused Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on animation paused outside active interval it is still paused Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on an idle animation without a start time it is still idle Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on an idle animation with a start time it is running assert_equals: expected "idle" but got "running"
FAIL After setting timeline on an idle animation with a sufficiently ancient start time it is finished assert_equals: expected "idle" but got "finished"
FAIL After setting timeline on a play-pending animation it begins playing after pending assert_true: Animation is initially play-pending expected true got undefined
FAIL After setting timeline on a pause-pending animation it becomes paused after pending promise_test: Unhandled rejection with value: object "TypeError: Cannot assign to read only property 'timeline' of object '#<Animation>'"
FAIL After clearing timeline on paused animation it is still paused assert_false: expected false got undefined
FAIL After clearing timeline on finished animation it is idle Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on running animation it is idle Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on idle animation it is still idle Cannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on play-pending animation it is still pending assert_true: expected true got undefined
FAIL After clearing and re-setting timeline on play-pending animation it begins to play assert_true: expected true got undefined
FAIL After clearing timeline on a pause-pending animation it is still pending assert_true: expected true got undefined
FAIL After clearing and re-setting timeline on a pause-pending animation it completes pausing assert_true: expected true got undefined
FAIL After clearing and re-setting timeline on an animation in the middle of an aborted pause, it continues playing using the same start time promise_test: Unhandled rejection with value: object "TypeError: Cannot assign to read only property 'timeline' of object '#<Animation>'"
Harness: the test ran to completion.
